voice-commands-api/requirements.txt
flask==3.0.3
SpeechRecognition==3.10.3
psycopg2-binary==2.9.9
scikit-learn==1.3.2
gTTS==2.5.1
pydub==0.25.1
